+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 2.2.0
 
- - Add ability to define multiple versions using one block
+ - Add ability to define multiple versions using one block.
+ - Added ability to mark certain attributes as optional when validating with `loose: true` and required otherwise.
 
 ## 2.1.1